# lib/api-zod/src/generated/types/apkInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApkInfo {
  appName: string;
  packageName: string;
  versionName?: string;
  versionCode?: string;
  urls: string[];
  imageCount: number;
}