# lib/api-zod/src/generated/types/imageInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageInfoType } from "./imageInfoType";

export interface ImageInfo {
  path: string;
  name: string;
  directory: string;
  size: number;
  type: ImageInfoType;
}