# lib/api-zod/src/generated/types/imageList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageInfo } from "./imageInfo";

export interface ImageList {
  images: ImageInfo[];
  total: number;
}